+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- feat(changelog): Add "auto" changeset policy
+- feat(changelog): Add "auto" changeset policy (#93)
+- fix(github): Ignore pending response from legacy commit check API (#94)
 
 ## 0.9.6