+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat(changelog): Add "auto" changeset policy (#93)
 - fix(github): Ignore pending response from legacy commit check API (#94)
+- feat(target): Add "docker" target (#95)
 
 ## 0.9.6